kernel: misc header and helper compat for 5.4
- input-event-codes.h / minmax.h only exist since 5.10 on the GKI
  trees tiann builds; fall back to linux/input.h / linux/kernel.h
  below (all upstream targets keep their includes verbatim),
- cast 5.4's void * ns_get_path() result for the long comparison,
- skip security_inode_init_security_anon() below 5.12 (verified
  absent on 5.4); the wrapper inode is labeled afterwards anyway.

+++ kernel/infra/su_mount_ns.c
@@ -68,7 +68,9 @@ try_setns:
         goto out;
     }
     struct path ns_path;
-    long ret = ns_get_path(&ns_path, pid1_task, &mntns_operations);
+    // 5.9+ returns int here; 5.4 returns void * (NULL on success), so cast
+    // to keep both warning-free. NULL/0 means success either way.
+    long ret = (long)ns_get_path(&ns_path, pid1_task, &mntns_operations);
     put_task_struct(pid1_task);
     if (ret) {
         pr_warn("failed get path for init mount namespace: %ld\n", ret);
